+++ aquasec/auth.py
@@ -106,7 +106,7 @@ def api_auth(api_key, api_secret, api_endpoint, api_role, api_methods, verbose=F
         "X-Signature": signature
     }
 
-    response = requests.post(api_url, headers=headers, data=post_body_str)
+    response = requests.post(api_url, headers=headers, data=post_body_str, timeout=10)
 
     # Extract status and token from the response
     if response.status_code == 200:
@@ -130,7 +130,7 @@ def user_pass_saas_auth(user, passwd, api_endpoint, verbose=False):
     if verbose:
         print(f"Auth URL: {api_url}")
     
-    res = requests.post(url=api_url, json=auth_info, verify=False)
+    res = requests.post(url=api_url, json=auth_info, verify=False, timeout=10)
 
     # One of the authentications above succeeded
     if res.status_code == 200:
@@ -162,7 +162,7 @@ def user_pass_onprem_auth(user, passwd, csp_endpoint):
 
     # Do authentication
     api_url = csp_endpoint + "/api/v1/login"
-    res = requests.post(url=api_url, json=auth_info, verify=False)
+    res = requests.post(url=api_url, json=auth_info, verify=False, timeout=10)
 
     # One of the authentications above succeeded
     if res.status_code == 200: